Move model scripting class into own file
This also includes the various associated classes.
