+++ code/source_groups.cmake
@@ -931,6 +931,8 @@ set(file_root_scripting_api
 	scripting/api/gamestate.cpp
 	scripting/api/hudgauge.h
 	scripting/api/hudgauge.cpp
+	scripting/api/model.h
+	scripting/api/model.cpp
 	scripting/api/vecmath.h
 	scripting/api/vecmath.cpp
 )
